requirements.txt
# DataNarrative Dependencies
# ==========================

# Core
pandas>=2.0.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.7.0
pillow>=10.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# AI (Optional - for enhanced features)
# Uncomment if you have API keys
# anthropic>=0.18.0

# Vector Store (Optional - for persistent storage)
# Uncomment for production use
# chromadb>=0.4.0
# sentence-transformers>=2.2.0

# Development
# pytest>=7.0.0
# httpx>=0.24.0
